requirements.txt
asgiref==3.5.2
Django==4.0.5
numpy==1.23.0
pip==22.1.2
setuptools==62.6.0
sqlparse==0.4.2
wheel==0.37.1
opencv-python==4.6.0.66
dlib~=19.24.0
mediapipe~=0.8.10.1
Pillow~=9.1.1